Update dependencies
This commit updates the build to install PyInstaller via pip
(`six8/pyinstaller-alpine` hasn't been updated in 4 years). As a result of this
change, the final binary now depends on glibc instead of musl, so we're also
switching to a glibc-based distribution for the runtime image.

Since PyInstaller now shares the venv with bagit-python, it appears to be able
to locate and bundle setuptools. I believe setuptools is installed because it's
a dependency of setuptools_scm which is listed in the setup.py. This means we
no longer need to manually install setuptools.

I've also taken this opportunity to upgrade Python to 3.13, as 3.8 reached its
end of life (EOL) a few days ago.